backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Async HTTP
httpx==0.27.0

# Redis (async)
redis[asyncio]==5.0.4

# PostgreSQL
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30

# Email parsing & DNS
dnspython==2.6.1
checkdmarc==4.5.1
dkimpy==1.1.5
python-whois==0.9.4

# URL / Domain analysis
rapidfuzz==3.9.3
confusable_homoglyphs==3.3.0
tldextract==5.1.2

# TLS / Crypto
cryptography==42.0.7

# HTML / Content parsing
beautifulsoup4==4.12.3
lxml==5.2.1

# VirusTotal
vt-py==0.18.0

# Playwright (headless browser)
playwright==1.44.0

# Rate limiting
slowapi==0.1.9

# Config / Env
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Utilities
structlog==24.1.0
tenacity==8.3.0
